Fix createdLogIndex overflow: integer → bigint in schema
Post-merge mainnet RPCs return synthetic MEV logs with sentinel
logIndex values near 0xFFFFFFFF (e.g. 0xFFFFFFFE = 4294967294).
These overflow PostgreSQL/PGlite integer (signed 32-bit max 2^31-1).
Changed createdLogIndex from p.integer() to p.bigint() on all three
tables (template, question, response) and cast event.log.logIndex
to BigInt in the insert handlers. Requires rm -rf .ponder/pglite.

+++ packages/ponder/ponder.schema.ts
@@ -9,7 +9,7 @@ export const template = onchainTable("template", (p) => ({
   user: p.hex().notNull(),
   questionText: p.text(),
   createdBlock: p.bigint().notNull(),
-  createdLogIndex: p.integer().notNull(),
+  createdLogIndex: p.bigint().notNull(),
   createdTxHash: p.hex().notNull(),
   createdTimestamp: p.bigint().notNull(),
 }));
@@ -65,7 +65,7 @@ export const question = onchainTable(
     reopensQuestionId: p.text(),
 
     createdBlock: p.bigint().notNull(),
-    createdLogIndex: p.integer().notNull(),
+    createdLogIndex: p.bigint().notNull(),
     createdTxHash: p.hex().notNull(),
     createdTimestamp: p.bigint().notNull(),
     updatedBlock: p.bigint().notNull(),
@@ -98,7 +98,7 @@ export const response = onchainTable("response", (p) => ({
 
   timestamp: p.bigint().notNull(),
   createdBlock: p.bigint().notNull(),
-  createdLogIndex: p.integer().notNull(),
+  createdLogIndex: p.bigint().notNull(),
   createdTxHash: p.hex().notNull(),
   revealedBlock: p.bigint(),
 }));
